requirements-test.txt
coverage==4.4.2
weather-api